Изменения react-burger.
