+++ src/hooks/useRegistration.js
@@ -1,31 +1,49 @@
-import { useCallback } from 'react';
+import { useState, useEffect, useCallback } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { registrationThunk } from '../services/actions/registration';
 import { getRegistrationSelector } from '../services/selectors/selector';
+import { RESET_STATE } from '../services/actions/auth';
 
 export const useRegistration = () => {
-  const { user, isSuccess, accessToken, refreshToken } = useSelector(
+  const { registrationRequest, registrationFailed } = useSelector(
     getRegistrationSelector
   );
 
+  const [isRegistered, setIsRegistered] = useState(false);
+  const [isLoading, setIsLoading] = useState(false);
+
   const dispatch = useDispatch();
 
+  // Регистрация пользователя.
   const registration = useCallback(
     (form) => {
-      dispatch(registrationThunk(form));
+      setIsLoading(true);
+      try {
+        dispatch(registrationThunk(form));
+      } catch (error) {
+      } finally {
+        setIsLoading(false);
+      }
     },
     [dispatch]
   );
 
-  const login = useCallback((form) => {}, []);
-
-  const logout = useCallback(() => {}, []);
+  // Проверяем результаты запроса регистрации пользователя.
+  useEffect(() => {
+    if (!isLoading && !registrationRequest && !registrationFailed) {
+      // Если был сделан запрос registration и он завершен успешно
+      // то в состоянии устанавливаем, что пользователь имеет статус зарегистрирован.
+      setIsRegistered(true);
+      // И сбрасываем все состояния в редукторе до первоначального.
+      dispatch({
+        type: RESET_STATE,
+      });
+    }
+  }, [dispatch, isLoading, registrationRequest, registrationFailed]);
 
   return {
-    user,
-    isSuccess,
+    isLoading,
+    isRegistered,
     registration,
-    accessToken,
-    refreshToken,
   };
 };
