+++ src/services/actions/forgot-password.js
@@ -4,6 +4,8 @@ export const POST_FORGOT_PASSWORD_FAILED = 'POST_FORGOT_PASSWORD_FAILED';
 export const POST_FORGOT_PASSWORD_SUCCESS = 'POST_FORGOT_PASSWORD_SUCCESS';
 export const POST_FORGOT_PASSWORD_REQUEST = 'POST_FORGOT_PASSWORD_REQUEST';
 
+export const RESET_STATE = 'RESET_STATE';
+
 // thunk
 export function forgotPasswordThunk(form) {
   return function (dispatch) {
@@ -12,14 +14,6 @@ export function forgotPasswordThunk(form) {
     });
     forgotPassword(form)
       .then((data) => {
-        // Проверяем внутренний статус ответа.
-        if (!data.success) {
-          dispatch({
-            type: POST_FORGOT_PASSWORD_FAILED,
-          });
-          return;
-        }
-
         dispatch({
           type: POST_FORGOT_PASSWORD_SUCCESS,
           data: data,
