+++ src/services/actions/password-reset.js
@@ -4,6 +4,8 @@ export const POST_PASSWORD_RESET_FAILED = 'POST_PASSWORD_RESET_FAILED';
 export const POST_PASSWORD_RESET_SUCCESS = 'POST_PASSWORD_RESET_SUCCESS';
 export const POST_PASSWORD_RESET_REQUEST = 'POST_PASSWORD_RESET_REQUEST';
 
+export const RESET_STATE = 'RESET_STATE';
+
 // thunk
 export function passwordResetThunk(form) {
   return function (dispatch) {
@@ -12,14 +14,6 @@ export function passwordResetThunk(form) {
     });
     passwordReset(form)
       .then((data) => {
-        // Проверяем внутренний статус ответа.
-        if (!data.success) {
-          dispatch({
-            type: POST_PASSWORD_RESET_FAILED,
-          });
-          return;
-        }
-
         dispatch({
           type: POST_PASSWORD_RESET_SUCCESS,
           data: data,
