--- src/services/reducers/logout.js
@@ -1,45 +1,0 @@
-import {
-  POST_LOGOUT_FAILED,
-  POST_LOGOUT_SUCCESS,
-  POST_LOGOUT_REQUEST,
-  RESET_STATE,
-} from '../actions/logout';
-
-const initialState = {
-  logoutRequest: true,
-  logoutFailed: false,
-};
-
-export const logoutReducer = (state = initialState, action) => {
-  switch (action.type) {
-    case POST_LOGOUT_REQUEST: {
-      return {
-        ...state,
-        logoutRequest: true,
-        logoutFailed: false,
-      };
-    }
-    case POST_LOGOUT_SUCCESS: {
-      return {
-        ...state,
-        logoutRequest: false,
-        logoutFailed: false,
-      };
-    }
-    case POST_LOGOUT_FAILED: {
-      return {
-        ...state,
-        logoutFailed: true,
-        logoutRequest: false,
-      };
-    }
-    case RESET_STATE: {
-      return {
-        ...initialState,
-      };
-    }
-    default: {
-      return state;
-    }
-  }
-};
