+++ src/services/reducers/rootReducer.js
@@ -4,9 +4,6 @@ import { orderReducer } from './order';
 import { modalReducer } from './modal';
 import { ingredientsConstructorReducer } from './ingredientsConstructor';
 import { registrationReducer } from './registration';
-import { loginReducer } from './login';
-import { refreshReducer } from './refresh';
-import { logoutReducer } from './logout';
 import { userReducer } from './user';
 import { forgotPasswordReducer } from './forgot-password';
 import { passwordResetReducer } from './password-reset';
@@ -18,9 +15,6 @@ export const rootReducer = combineReducers({
   order: orderReducer,
   modal: modalReducer,
   registration: registrationReducer,
-  login: loginReducer,
-  refresh: refreshReducer,
-  logout: logoutReducer,
   user: userReducer,
   forgotPassword: forgotPasswordReducer,
   passwordReset: passwordResetReducer,
