+++ src/services/selectors/selector.js
@@ -9,12 +9,6 @@ export const getModalSelector = (store) => store.modal;
 
 export const getRegistrationSelector = (store) => store.registration;
 
-export const getLoginSelector = (store) => store.login;
-
-export const getRefreshTokenSelector = (store) => store.refresh;
-
-export const getLogoutSelector = (store) => store.logout;
-
 export const userSelector = (store) => store.user;
 
 export const getForgotPasswordSelector = (store) => store.forgotPassword;
