+++ src/utils/burger-api.js
@@ -85,7 +85,7 @@ export function logout(refreshToken) {
  * @returns Promise<data>
  */
 export function getUser(token) {
-  return fetch(`${NORMA_API}/auth/user`, {
+  return fetch(`${NORMA_API}/auth/user1`, {
     method: 'GET',
     headers: {
       'Content-Type': 'application/json',
@@ -143,5 +143,12 @@ export function passwordReset(data) {
  * @returns Promise<json> || Promise<err>
  */
 const checkResponse = (res) => {
-  return res.ok ? res.json() : res.json().then((err) => Promise.reject(err));
+  // return res.ok ? res.json() : res.json().then((err) => Promise.reject(err));
+  // return res.ok ? res.json() : res.json().then((err) => Promise.reject(err));
+
+  if (res.ok) {
+    return res.json();
+  }
+
+  throw new Error('401');
 };
